Update to Prost v11 and Tonic v9 [1/3]
This patch contains the dependency updates and cargo.toml changes for
updating the prost and tonic versions to their respective latest.

The biggest change in this version is that Prost no longer bundles the
protoc compiler or the well known protos which leaves us to update our
tooling to handle them.