Add common_proto folder [2/3]
This patch adds the common proto folder for common google protos which
are not included by default in all of the repos who's protos we are
consuming because prost previously bunlded them and they are generally
bundled with protoc installations.

This patch includes those common protos in this repo for ease of
reproducibility.