Always compile CosmosSDK staking module protos, add disclaimer
As mentioned in the README.md, the staking module can be particularly
difficult to work with since it always compiles to have a namespace
conflict. The README mentions that file should usually be ignored due to
disuse of the conflicting values, with an explanation of the manual work
to fix the conflict.